refactor: drop pass-through schema Column wrapper; re-export the domain type
The user-facing schema.Column had the same four fields, types, and defaults as the
domain Column, and to_domain_column copied them across unchanged — a shallow module
whose interface was exactly as complex as what it wrapped, plus a sync point: every
field added to the domain Column had to be mirrored here and in the conversion.

Re-export the domain Column from the schema package and drop the wrapper. DeltaTable
stores domain Columns directly and to_desired_table passes them through. Callers doing
`from delta_engine.adapters.schema import Column` are unaffected — they get the real
type. Delete the wrapper file and its dedicated test.
